+++ Modules/Filtering/TextureFeatures/test/RunLengthTextureFeaturesImageFilterTestWithoutMask.cxx
@@ -45,7 +45,7 @@ RunLengthTextureFeaturesImageFilterTestWithoutMask(int argc, char * argv[])
   const unsigned int VectorComponentDimension = 10;
 
   // Declare types
-  typedef int                                                             InputPixelType;
+  typedef float                                                           InputPixelType;
   typedef float                                                           OutputPixelComponentType;
   typedef itk::Vector<OutputPixelComponentType, VectorComponentDimension> OutputPixelType;
